clippy fixes

+++ dash/src/blockdata/transaction/txin.rs
@@ -102,6 +102,6 @@ mod tests {
         assert_eq!(txin.script_sig, ScriptBuf::new());
         assert_eq!(txin.sequence, 0xFFFFFFFF);
         assert_eq!(txin.previous_output, OutPoint::default());
-        assert_eq!(txin.witness.len(), 0 as usize);
+        assert_eq!(txin.witness.len(), 0_usize);
     }
 }
